+++ src/sqlalchemy_cratedb/support/__init__.py
@@ -4,12 +4,13 @@ from sqlalchemy_cratedb.support.polyfill import (
     patch_autoincrement_timestamp,
     refresh_after_dml,
 )
-from sqlalchemy_cratedb.support.util import refresh_dirty, refresh_table
+from sqlalchemy_cratedb.support.util import quote_relation_name, refresh_dirty, refresh_table
 
 __all__ = [
     check_uniqueness_factory,
     insert_bulk,
     patch_autoincrement_timestamp,
+    quote_relation_name,
     refresh_after_dml,
     refresh_dirty,
     refresh_table,